Export results of a=fmtp TROFF, CMAX and PAR fields
These fields are parsed by the SMPTE ST2110-20 parser but it does not
export (copy) them to the resulting smpte2110_fmtp structure.

Changed to copy these values into the smpte2110_fmtp structure.

+++ smpte2110_sdp_parser.h
@@ -140,7 +140,7 @@ struct smpte2110_media_attr_fmtp_params {
 	uint16_t maxudp;
 	struct smpte_2110_par par;
 	uint64_t troff;
-	int cmax;
+	uint16_t cmax;
 };
 
 struct smpte2110_media_attr_fmtp {
